+++ src/speakers/controllers/speakers.controller.ts
@@ -1,5 +1,4 @@
 import { Controller, Headers, Get, Post, Body, Res, Param, Put, Delete, UseGuards, HttpStatus, HttpException } from '@nestjs/common';
-import { AuthGuard } from '@nestjs/passport';
 import { Response } from 'express';
 import { LocalStrategy } from 'src/auth/local.auth';
 import { MiddlewareApp } from 'src/shared/middleware/middleware.middleware';
@@ -12,8 +11,13 @@ export class SpeakerController {
 
   @Get(':id')
   @UseGuards(MiddlewareApp)
-  async findOne(@Param("id") speakerID: string, @Res() res: Response): Promise<object>{
+  async findOne(@Headers('authorization') authHeader: string, @Param("id") speakerID: string, @Res() res: Response): Promise<object>{
     try{
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
+
       const speaker = await this.speakerService.findOne(speakerID);
       if (!speaker) return res.status(HttpStatus.BAD_REQUEST).json({message: "Speaker not found"});
       return res.status(200).json({"message": speaker});
@@ -23,18 +27,13 @@ export class SpeakerController {
     }
   }
 
-  // @Get()
-  // async findAll(@Header() authorization: string): Promise<Speaker[]> {
-  //   console.log(req.authorization)
-  //   return this.speakerService.findAll();
-  // }
-
   @Get()
   async findAll(@Headers('authorization') authHeader: string,  @Res() res: Response): Promise<Speaker[] | object> {
     try{
       let token = authHeader.split(' ')[1];
       let userFound = await this.authService.validate(token);
 
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       let speakers = await this.speakerService.findAll();
       return res.status(HttpStatus.OK).json({speakers});
     }catch{
@@ -44,10 +43,14 @@ export class SpeakerController {
 
   @Post()
     @UseGuards(MiddlewareApp)
-    async create(@Body() speaker: Speaker, @Res() res: Response): Promise<Speaker | object> {
+    async create(@Headers('authorization') authHeader: string, @Body() speaker: Speaker, @Res() res: Response): Promise<Speaker | object> {
     try {
-      const createdSpeaker = await this.speakerService.create(speaker);
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       
+      const createdSpeaker = await this.speakerService.create(speaker);
       if(!createdSpeaker) res.status(HttpStatus.BAD_REQUEST).json({message: "Speaker already created"});
       return res.status(HttpStatus.CREATED).json(createdSpeaker);
     } catch (error) {
@@ -57,10 +60,14 @@ export class SpeakerController {
 
   @Put(':id')
   @UseGuards(MiddlewareApp)
-  async update(@Param("id") speakerID: string, @Body() speaker: Speaker, @Res() res: Response): Promise<object> {
+  async update(@Headers('authorization') authHeader: string, @Param("id") speakerID: string, @Body() speaker: Speaker, @Res() res: Response): Promise<object> {
     try{
-      let result = await this.speakerService.update(speakerID, speaker);
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       
+      let result = await this.speakerService.update(speakerID, speaker);
       if(!result) return res.status(HttpStatus.BAD_REQUEST).json({message: "Speaker failed to be updated"})
       return res.status(200).json({message: "Speaker updated successfully"})
     }catch(err){
@@ -70,10 +77,14 @@ export class SpeakerController {
 
   @Delete(':id')
   @UseGuards(MiddlewareApp)
-  async delete(@Param("id") speakerID: string,  @Res() res: Response): Promise<object> {
+  async delete(@Headers('authorization') authHeader: string, @Param("id") speakerID: string,  @Res() res: Response): Promise<object> {
     try{
-      let result = await this.speakerService.delete(speakerID);
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
 
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
+      
+      let result = await this.speakerService.delete(speakerID);
       if(!result) return res.status(HttpStatus.BAD_REQUEST).json({message: "Speaker failed to be deleted"})
       return res.status(200).json({message: "Speaker deleted successfully"})
     }catch(err){
