+++ src/users/controllers/users.controller.ts
@@ -1,21 +1,26 @@
-import { Controller, Get, Post, Put, Res, Delete, Param, Body, UseGuards, HttpStatus, HttpException } from '@nestjs/common';
+import { Controller, Headers, Get, Post, Put, Res, Delete, Param, Body, UseGuards, HttpStatus, HttpException } from '@nestjs/common';
 import { UsersService } from '../services/users.service';
 import { User } from '../../shared/models/user.schema';
 import { MiddlewareApp } from 'src/shared/middleware/middleware.middleware';
 import { Response } from 'express';
+import { LocalStrategy } from 'src/auth/local.auth';
 
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(private readonly usersService: UsersService, private authService: LocalStrategy) {}
 
   @Get(':id')
   @UseGuards(MiddlewareApp)
-  async findOne(@Param("id") userID: string, @Res() res: Response): Promise<object> {
+  async findOne(@Headers('authorization') authHeader: string, @Param("id") userID: string, @Res() res: Response): Promise<object> {
     try{
-      const participant = await this.usersService.findOne(userID);
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
 
-      if (!participant) return res.status(HttpStatus.BAD_REQUEST).json({message: "User not found"});
-      return res.status(200).json({"message": participant});
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
+      const user = await this.usersService.findOne(userID);
+
+      if (!user) return res.status(HttpStatus.BAD_REQUEST).json({message: "User not found"});
+      return res.status(200).json({"message": user});
     }catch(err){
       console.error(err);
       return res.status(500).json({message: "Internal server error"});
@@ -24,14 +29,27 @@ export class UsersController {
 
   @Get()
   @UseGuards(MiddlewareApp)
-  async findAll(): Promise<User[]> {
-    return this.usersService.findAll();
+  async findAll(@Headers('authorization') authHeader: string, @Res() res: Response): Promise<object> {
+    try{
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
+      return res.status(HttpStatus.ACCEPTED).json({message: this.usersService.findAll()});
+    }catch(err){
+      console.error(err);
+      return res.status(500).json({message: "Internal server error"});
+    }
   }
 
   @Post()
   @UseGuards(MiddlewareApp)
-  async create(@Body() user: User, @Res() res: Response): Promise<object> {
+  async create(@Headers('authorization') authHeader: string, @Body() user: User, @Res() res: Response): Promise<object> {
     try {
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       const createdUser = await this.usersService.create(user);;
       
       if(!createdUser) res.status(HttpStatus.BAD_REQUEST).json({message: "User already created"});
@@ -43,8 +61,12 @@ export class UsersController {
 
   @Put(':id')
   @UseGuards(MiddlewareApp)
-  async update(@Param("id") userID: string, @Body() user: User, @Res() res: Response): Promise<object> {
+  async update(@Headers('authorization') authHeader: string, @Param("id") userID: string, @Body() user: User, @Res() res: Response): Promise<object> {
     try{
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       let result = await this.usersService.update(userID, user);
       
       if(!result) return res.status(HttpStatus.BAD_REQUEST).json({message: "User failed to be updated"})
@@ -56,8 +78,12 @@ export class UsersController {
 
   @Delete(':id')
   @UseGuards(MiddlewareApp)
-  async delete(@Param("id") userID: string,  @Res() res: Response): Promise<object> {
+  async delete(@Headers('authorization') authHeader: string, @Param("id") userID: string,  @Res() res: Response): Promise<object> {
     try{
+      let token = authHeader.split(' ')[1];
+      let userFound = await this.authService.validate(token);
+
+      if (!userFound) return res.status(HttpStatus.UNAUTHORIZED).json({message: "Unauthorized"});
       let result =await this.usersService.delete(userID);
 
       if(!result) return res.status(HttpStatus.BAD_REQUEST).json({message: "User failed to be deleted"})
